pyproject.toml
[project]
name = "book-rag-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.9",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.59",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.16",
    "openai>=1.78.1",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.6",
    "pymupdf>=1.25.5",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "unstructured>=0.17.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
